requirements.txt
beautifulsoup4
deep-translator
fastapi
lxml
Pillow
redis
uvicorn[standard] 
aiofiles
apscheduler
httpx[http2]
honcho
./src/wheels/fastapi_limiter-0.1.5-py3-none-any.whl